\name{breastfeed}
\alias{breastfeed}
\docType{data}
\title{ Breast feeding decision}
\description{
  In a UK hospital, 135 expectant mothers were surveyed on the decision of breastfeeding their
babies or not, along with two-level predictive factors 
}
\usage{data(breastfeed)}

\source{
Stephane Heritier, Eva Cantoni, Samuel Copt and Maria-Pia Victoria-Fese (2009). \emph{Robust Methods in Biostatistics}, John Wiley & Sons}

\examples{
data(breastfeed)
str(breastfeed)
}
\keyword{datasets}
