\name{docvisits}
\alias{docvisits}
\docType{data}
\title{ Doctor visits}
\description{
  A cohort of 3066 Americans over the age of 50 were studied on health care utilization, doctor
office visits.
}
\usage{data(docvisits)}

\source{
Heritier, Stephane and Cantoni, Eva and Copt, Samuel and Victoria-Feser, Maria-Pia (2009). \emph{Robust Methods in Biostatistics}, John Wiley \& Sons}

\examples{
data(docvisits)
str(docvisits)
}
\keyword{datasets}
