% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mpath.R
\name{find_optimal_cluster_number}
\alias{find_optimal_cluster_number}
\title{find_optimal_cluster_number identifies the optimal number of initial cluster number by searching from min_cluster_num to max_cluster_num}
\usage{
find_optimal_cluster_number(rpkmFile, sampleFile, min_cluster_num = 7,
  max_cluster_num = 13, diversity_cut = 0.6, size_cut = 0.05)
}
\arguments{
\item{rpkmFile:}{a tab delimited txt file of expression data, containing cells in columns and genes in rows}

\item{sampleFile:}{a tab delimited txt file of sample annotation with two columns, the first column is cell ID, the second column is group ID}

\item{min_cluster_num:}{minimum number of initial clusters}

\item{max_cluster_num:}{maximum number of initial clusters}

\item{diversity_cut:}{the cutoff value of diversity for differentiating landmark clusters from non-landmark clusters. The diversity of a landmark cluster must be below this cutoff.}

\item{size_cut:}{the cutoff value of size i.e. number of cells for differentiating landmark clusters from non-landmark clusters. The number of cells in a landmark cluster must be greater than this cutoff.}
}
\description{
find_optimal_cluster_number identifies the optimal number of initial cluster number by searching from min_cluster_num to max_cluster_num
}
\examples{
rpkmFile = "TPM_GSE60783_noOutlier_geneQC0.05anyGroup_CD4vsCD8DEG.txt";
sampleFile = "sample_GSE60783_noOutlier.txt";
find_optimal_cluster_number(rpkmFile = rpkmFile,
                            sampleFile = sampleFile,
                            min_cluster_num = 7, max_cluster_num = 15,
                            diversity_cut = 0.6, size_cut = 0.05)
}

