% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mpath.R
\name{heatmap_nbor}
\alias{heatmap_nbor}
\title{heatmap_nbor plot heatmap of gene expression}
\usage{
heatmap_nbor(exprs, cell_order, plot_genes, cell_annotation,
  num_gene_cluster = 4, hm_height = 10, hm_width = 10, baseName,
  colorPalette = NULL, n_linechart = NULL)
}
\arguments{
\item{exprs}{a data frame or matrix of expression data(ie. rpkm, TPM, fpkm) containing cells in columns and genes in rows}

\item{cell_order}{a vector storing the order of cells with cell ID or name, same as appeared in column names of \code{exprs}}

\item{plot_genes}{a vector storing the genes selected for plot, same as appeared in the row names of \code{exprs}}

\item{cell_annotation}{a two column data frame or matrix annotating cells(cell ID or name) with cell types}

\item{num_gene_cluster}{a integer indicating the number of gene clusters to generate by cutting the dendrogram tree, if num_gene_cluster = NULL, no gene clusters will be generated}

\item{hm_height}{an integeter to specify the heatmap height}

\item{hm_width}{an integeter to specify heatmap width}

\item{baseName}{a character string to specify the prefix of name of result files}

\item{colorPalette}{a character vector to specify the color scheme of column side bar that labels cell type of individual cells. Default value is NULL, a default color scheme will be deployed. Alternatively users can specfiy their desired color paletter, for examples, colorPalette=c("red","green","blue","black","orange","purple","burlywood4")}

\item{n_linechart}{a string vector to specify for which cell ordering to plot the line chart. Default value is NULL, all the cells in cell_order will be plot. Alternatively, users could specify, for examples two line charts, by n_linechart = list(c("cell1","cell2"),c("cell3","cell4"))}
}
\description{
heatmap_nbor plot heatmap of gene expression
}
\examples{
heatmap_nbor(exprs = "TPM_GSE60783_noOutlier_geneQC0.05anyGroup.txt",                        
             cell_order = "CDP_2_CDP_1_preDC_9_preDC_3_order.txt", 
             plot_genes = "CDP_2_CDP_1_preDC_9_preDC_3_vgam_deg0.05.txt",
             cell_annotation = "sample_GSE60783_noOutlier.txt", 
             num_gene_cluster = 6,
             hm_height = 15, hm_width = 10,
             baseName = "CDP_2_CDP_1_preDC_9_preDC_3_order_vgam_deg0.05") 

}

