% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QC_gene.R
\name{QC_gene}
\alias{QC_gene}
\title{QC_gene removes genes that have TPM values < 1 in more than 95 percent of cells in each group}
\usage{
QC_gene(rpkmFile = "TPM_monocyte_Mar2015_noOutlier.txt",
  rpkmQCFile = "TPM_monocyte_Mar2015_noOutlier_geneQC0.05perGroup.txt",
  sampleFile = "monocyte_sample_Mar2015_noOutlier.txt", threshold = 0.05,
  method = "any")
}
\arguments{
\item{rpkmFile}{a tab delimited file containing expression data (TPM, CPM, FPKM, etc), columns are cells and rows are genes.}

\item{rpkmQCFile}{resulting file after QC}

\item{sampleFile}{a tab delimited file containing sample annotations with two columns. The first column indicates SampleID, the second column indicates GroupID.}

\item{threshold}{the cutoff of percentage of cells in which the given gene is not expressed.Default is 0.05.}

\item{method}{keep genes whose rpkm values are not less than 1 in at least 5 percent of the cells in every group, method="any": keep genes whose rpkm values are not less than 1 in at least 5 percent of the cells in any group. Default is 'any'.}
}
\description{
QC_gene removes genes that have TPM values < 1 in more than 95 percent of cells in each group
}
\examples{
\dontrun{
QC_gene(rpkmFile = "TPM_MDP_CDP_preDC_Mar2015_noOutlier.txt",
        rpkmQCFile = "TPM_MDP_CDP_preDC_Mar2015_noOutlier_geneQC0.05anyGroup.txt",
        sampleFile = "sample_MDP_CDP_preDC_Mar2015_noOutlier.txt",
        threshold = 0.05,method = "any")
}
}
