% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_hc.R
\name{SC_hc}
\alias{SC_hc}
\title{SC_hc hierachical clustering of single cells}
\usage{
SC_hc(dataFile, baseName, cuttree_k = 4)
}
\arguments{
\item{dataFile}{a tab delimited txt file of expression data, columns are cells, rows are genes.}

\item{baseName}{prefix name of resulting files}

\item{cuttree_k}{number of clusters to be generated by cutting the dendrogram.}
}
\description{
SC_hc hierachical clustering of single cells
}
