% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_hc.R
\name{SC_hc_colorCode}
\alias{SC_hc_colorCode}
\title{SC_hc_colorCode hierarchical clustering of single cells, leaves color-coded by cell type information.}
\usage{
SC_hc_colorCode(dataFile, baseName = NULL, cuttree_k = NULL,
  sampleFile = NULL, width = 22, height = 10, iflog2 = TRUE,
  colorPalette = NULL, distMethod = "Euclidean")
}
\arguments{
\item{dataFile:}{a tab delimited txt file of expression data, columns are cells, rows are genes.}

\item{baseName:}{prefix name of resulting files}

\item{cuttree_k:}{number of clusters to be generated by cutting the dendrogram. cuttree_k can be an integer or NULL. If NULL, no clusters will be generated.}

\item{sampleFile:}{a tab delimited txt file of cell annotation, the first column is cell ID, the second column is group ID.}

\item{width:}{width of the resulting figure}

\item{height:}{height of the resulting figure}

\item{iflog2:}{a boolean value indiciating if the expression data will be log2 transformed.}

\item{colorPalette:}{if NULL, brewer.pal(n,"Set2") will be used, or colorPalette can be c("red","green","blue").}

\item{distMethod:}{distMethod can be one of "pearson", "kendall", "spearman" or "Euclidean"}
}
\description{
SC_hc_colorCode hierarchical clustering of single cells, leaves color-coded by cell type information.
}
\examples{
dataFile = "TPM_GSE60783_noOutlier_geneQC0.05anyGroup_CD4vsCD8DEG.txt";
SC_hc_colorCode(dataFile = dataFile,
               cuttree_k = 11,
               sampleFile= "sample_GSE60783_noOutlier.txt",
               width = 22, height = 10, iflog2 = TRUE,
               colorPalette = c("red","green","blue"))
}

