% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mpath.R
\name{trim_net}
\alias{trim_net}
\title{trim_net trimms the weighted neighborhood network by removing edges of lower weights}
\usage{
trim_net(nb12, textSize = 20, baseName = NULL, method = "mst",
  start = "MDP_6", writeRes = TRUE)
}
\arguments{
\item{nb12}{a matrix of weighted neighborhood network, column and row names are landmarks, the values represent the weights of the edges connecting two landmarks}

\item{textSize}{the size of text}

\item{baseName}{a character string indicating the prefix name of resulting files}

\item{method}{trimming method, method can be one of "TrimNet" or "mst". When method="TrimNet" the initial node needs to be specified. Default is "mst"}

\item{start}{starting landmark, needs to be specified when method="TrimNet".}

\item{writeRes}{a boolean to indicate whether to save result files}
}
\value{
a matrix of trimmed state transition network, column and row names are landmarks, the values are 0 or 1 indicating whether the two landmarks are connected.
}
\description{
trim_net trimms the weighted neighborhood network by removing edges of lower weights
}
\examples{
\dontrun{
baseName = "TPM_GSE60783_noOutlier_geneQC0.05anyGroup_CD4vsCD8DEG";
trimmed_net <- trim_net(neighbor_network,textSize=30,
                        baseName = baseName,
                        method = "mst")
}
}
