% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mpath.R
\name{vgam_perGene}
\alias{vgam_perGene}
\title{vgam_perGene determines if a gene is differentially expressed along the re-ordered single-cell trajectories using vgam}
\usage{
vgam_perGene(expr, order, min_expr)
}
\arguments{
\item{expr:}{a vector of one gene's expression in different cells (ie. rpkm, TPM, fpkm)}

\item{order:}{a vector of re-ordered cell IDs}

\item{min_expr:}{a numeric value indicating the minimum TPM value for a gene to be considered as expressed. Default is 1.}
}
\value{
pval: p value of significance of the gene being differentially expressed
}
\description{
vgam_perGene determines if a gene is differentially expressed along the re-ordered single-cell trajectories using vgam
}
\examples{
p_val <- vgam_perGene(expr,order,min_expr=1)
}

